requirements.txt
# ===============================================================================
# AGENTE IA OYP 6.0 - DEPENDENCIAS PRINCIPALES
# ===============================================================================

# Framework Web Principal
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0

# Base de datos y ORM
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Utilidades Core
pydantic>=2.7.1,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
psutil==5.9.8

# Logging y Monitoreo
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Docker y deployment
docker==6.1.3

# ===============================================================================
# MÓDULO: Document Processor
# ===============================================================================
# Extracción de texto y metadatos
pymupdf==1.24.1
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.12.0
python-docx==1.1.0
docx2txt==0.8
unstructured==0.14.2
unstructured-inference==0.7.33
# Extras for unstructured - ensure all parsing capabilities
unstructured[local-inference]
unstructured[all-docs]


# Procesamiento de imágenes y OCR
pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
paddleocr==2.7.3
rapidocr-onnxruntime==1.3.19

# ===============================================================================
# MÓDULO: AI Engine
# ===============================================================================
# NLP y Machine Learning
torch==2.2.0
transformers==4.35.2
huggingface-hub==0.19.4
sentence-transformers==2.2.2
langdetect==1.0.9
rake-nltk==1.0.6
scikit-learn==1.3.2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langsmith==0.1.57

# ===============================================================================
# MÓDULO: Analytics Engine
# ===============================================================================
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
polars==0.19.19

# ===============================================================================
# MÓDULO: Report Generator
# ===============================================================================
weasyprint==60.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ===============================================================================
# MÓDULO: Gateway
# ===============================================================================
# (Usa dependencias core como FastAPI, Uvicorn, httpx)
# No hay dependencias adicionales específicas para el gateway aquí.
# ===============================================================================